+++ src/index.ts
@@ -145,7 +145,7 @@ export { useToggleLocalMute } from './hooks/sdk/useToggleLocalMute';
 export { useMeetingStatus } from './hooks/sdk/useMeetingStatus';
 export { useLocalAudioInputActivity } from './hooks/sdk/useLocalAudioInputActivity';
 export { useLocalAudioInputActivityPreview } from './hooks/sdk/useLocalAudioInputActivityPreview';
-export { useDevicePermissionStatus } from './hooks/sdk/useDevicePermissionStatus';
+export { useDeviceLabelTriggerStatus } from './hooks/sdk/useDeviceLabelTriggerStatus';
 export { useMediaStreamMetrics } from './hooks/sdk/useMediaStreamMetrics';
 
 // Providers
@@ -187,9 +187,9 @@ export { RosterAttendeeType, RosterType } from './types';
 // enums
 export {
   MeetingStatus,
-  DevicePermissionStatus,
   DeviceLabels,
   DeviceLabelTrigger,
+  DeviceLabelTriggerStatus,
 } from './types';
 export { Severity, ActionType } from './providers/NotificationProvider';
 
@@ -197,6 +197,7 @@ export { Severity, ActionType } from './providers/NotificationProvider';
 export { MeetingManager } from './providers/MeetingProvider/MeetingManager';
 
 // Interface
+export { MeetingManagerJoinOptions } from './providers/MeetingProvider/types';
 export { NotificationType, Action } from './providers/NotificationProvider';
 
 // Utilities