+++ src/providers/MeetingProvider/MeetingManager.ts
@@ -21,10 +21,14 @@ import {
   VideoInputDevice,
 } from 'amazon-chime-sdk-js';
 
-import { DeviceLabels, DeviceLabelTrigger, MeetingStatus } from '../../types';
+import {
+  DeviceLabels,
+  DeviceLabelTrigger,
+  DeviceLabelTriggerStatus,
+  MeetingStatus,
+} from '../../types';
 import {
   AttendeeResponse,
-  DevicePermissionStatus,
   FullDeviceInfoType,
   MeetingManagerJoinOptions,
   ParsedJoinParams,
@@ -71,18 +75,19 @@ export class MeetingManager implements AudioVideoObserver {
     device: VideoInputDevice | undefined
   ) => void)[] = [];
 
-  deviceLabelTriggerChangeObservers: (() => void)[] = [];
-
   audioInputDevices: MediaDeviceInfo[] | null = null;
 
   audioOutputDevices: MediaDeviceInfo[] | null = null;
 
   videoInputDevices: MediaDeviceInfo[] | null = null;
 
-  devicePermissionStatus = DevicePermissionStatus.UNSET;
+  deviceLabelTriggerStatus = DeviceLabelTriggerStatus.UNTRIGGERED;
 
-  devicePermissionsObservers: ((permission: DevicePermissionStatus) => void)[] =
-    [];
+  deviceLabelTriggerStatusObservers: ((
+    status: DeviceLabelTriggerStatus
+  ) => void)[] = [];
+
+  deviceLabelTriggerObservers: (() => void)[] = [];
 
   activeSpeakerListener: ((activeSpeakers: string[]) => void) | null = null;
 
@@ -320,8 +325,8 @@ export class MeetingManager implements AudioVideoObserver {
       }
 
       callback = async (): Promise<MediaStream> => {
-        this.devicePermissionStatus = DevicePermissionStatus.IN_PROGRESS;
-        this.publishDevicePermissionStatus();
+        this.deviceLabelTriggerStatus = DeviceLabelTriggerStatus.IN_PROGRESS;
+        this.publishDeviceLabelTriggerStatus();
         try {
           const devices = await navigator.mediaDevices.enumerateDevices();
           const hasVideoInput = devices.some(
@@ -333,13 +338,13 @@ export class MeetingManager implements AudioVideoObserver {
             video: constraints.video && hasVideoInput,
           });
 
-          this.devicePermissionStatus = DevicePermissionStatus.GRANTED;
-          this.publishDevicePermissionStatus();
+          this.deviceLabelTriggerStatus = DeviceLabelTriggerStatus.GRANTED;
+          this.publishDeviceLabelTriggerStatus();
           return stream;
         } catch (error) {
           console.error('MeetingManager failed to get device permissions');
-          this.devicePermissionStatus = DevicePermissionStatus.DENIED;
-          this.publishDevicePermissionStatus();
+          this.deviceLabelTriggerStatus = DeviceLabelTriggerStatus.DENIED;
+          this.publishDeviceLabelTriggerStatus();
           throw new Error(error);
         }
       };
@@ -507,7 +512,7 @@ export class MeetingManager implements AudioVideoObserver {
 
   invokeDeviceProvider = (deviceLabels: DeviceLabels): void => {
     this.setupDeviceLabelTrigger(deviceLabels);
-    this.publishDeviceLabelTriggerChange();
+    this.publishDeviceLabelTrigger();
   };
 
   /**
@@ -555,26 +560,6 @@ export class MeetingManager implements AudioVideoObserver {
     this.activeSpeakerCallbacks.forEach((callback) => {
       callback(this.activeSpeakers);
     });
-  };
-
-  subscribeToDevicePermissionStatus = (
-    callback: (permission: DevicePermissionStatus) => void
-  ): void => {
-    this.devicePermissionsObservers.push(callback);
-  };
-
-  unsubscribeFromDevicePermissionStatus = (
-    callbackToRemove: (permission: DevicePermissionStatus) => void
-  ): void => {
-    this.devicePermissionsObservers = this.devicePermissionsObservers.filter(
-      (callback) => callback !== callbackToRemove
-    );
-  };
-
-  private publishDevicePermissionStatus = (): void => {
-    for (const observer of this.devicePermissionsObservers) {
-      observer(this.devicePermissionStatus);
-    }
   };
 
   subscribeToSelectedVideoInputDevice = (
@@ -661,22 +646,40 @@ export class MeetingManager implements AudioVideoObserver {
     });
   };
 
-  subscribeToDeviceLabelTriggerChange = (callback: () => void): void => {
-    this.deviceLabelTriggerChangeObservers.push(callback);
+  subscribeToDeviceLabelTrigger = (callback: () => void): void => {
+    this.deviceLabelTriggerObservers.push(callback);
   };
 
-  unsubscribeFromDeviceLabelTriggerChange = (
-    callbackToRemove: () => void
+  unsubscribeFromDeviceLabelTrigger = (callbackToRemove: () => void): void => {
+    this.deviceLabelTriggerObservers = this.deviceLabelTriggerObservers.filter(
+      (callback) => callback !== callbackToRemove
+    );
+  };
+
+  private publishDeviceLabelTrigger = (): void => {
+    for (const callback of this.deviceLabelTriggerObservers) {
+      callback();
+    }
+  };
+
+  subscribeToDeviceLabelTriggerStatus = (
+    callback: (permission: DeviceLabelTriggerStatus) => void
   ): void => {
-    this.deviceLabelTriggerChangeObservers =
-      this.deviceLabelTriggerChangeObservers.filter(
+    this.deviceLabelTriggerStatusObservers.push(callback);
+  };
+
+  unsubscribeFromDeviceLabelTriggerStatus = (
+    callbackToRemove: (permission: DeviceLabelTriggerStatus) => void
+  ): void => {
+    this.deviceLabelTriggerStatusObservers =
+      this.deviceLabelTriggerStatusObservers.filter(
         (callback) => callback !== callbackToRemove
       );
   };
 
-  private publishDeviceLabelTriggerChange = (): void => {
-    for (const callback of this.deviceLabelTriggerChangeObservers) {
-      callback();
+  private publishDeviceLabelTriggerStatus = (): void => {
+    for (const observer of this.deviceLabelTriggerStatusObservers) {
+      observer(this.deviceLabelTriggerStatus);
     }
   };